+++ CHANGELOG.md
@@ -28,6 +28,10 @@
   no reason to read it when writing components.
 * **\[arez\]** Fix invariant failure that occurred when generating debug state for a component that extends
   `ReactArezComponent` that has observed an Arez observable that has no accessor defined.
+* **\[core\]** Remove invariant checking based on the state of the react component. This invariant checking
+  was duplicating functionality already implemented by the underlying `ReactJS` library. This also involved
+  removing the compile time setting `react4j.check_component_state_invariants` and the accessor
+  `ReactConfig.checkComponentStateInvariants()`.
 
 ### [v0.108](https://github.com/react4j/react4j/tree/v0.108) (2018-11-22)
 [Full Changelog](https://github.com/react4j/react4j/compare/v0.107...v0.108)